finish the renaming
