+++ drivers/native-driver/src/nativeMain/kotlin/com/squareup/sqldelight/drivers/native/SqliterSqlCursor.kt
@@ -1,4 +1,4 @@
-package com.squareup.sqldelight.drivers.ios
+package com.squareup.sqldelight.drivers.native
 
 import co.touchlab.sqliter.Cursor
 import co.touchlab.sqliter.getBytesOrNull
