+++ settings.gradle
@@ -1,7 +1,7 @@
 rootProject.name = 'sqldelight'
 
 include ':drivers:android-driver'
-include ':drivers:ios-driver'
+include ':drivers:native-driver'
 include ':drivers:sqlite-driver'
 include ':drivers:driver-test'
